Remove independent canisterdataprovider lib
