+++ rs/registry/canister-client/BUILD.bazel
@@ -5,8 +5,14 @@ package(default_visibility = ["//visibility:public"])
 DEPENDENCIES = [
     # Keep sorted.
     "//rs/interfaces/registry",
+    "//rs/nns/common",
+    "//rs/nns/constants",
+    "//rs/registry/transport",
     "//rs/types/types",
+    "@crate_index//:candid",
     "@crate_index//:ic-cdk",
+    "@crate_index//:ic-stable-structures",
+    "@crate_index//:itertools",
 ]
 
 DEV_DEPENDENCIES = [
