+++ rs/registry/canister-client/Cargo.toml
@@ -8,8 +8,16 @@ documentation.workspace = true
 
 [dependencies]
 ic-interfaces-registry = { path = "../../interfaces/registry" }
+ic-registry-transport = { path = "../transport" }
 ic-types = { path = "../../types/types" }
+ic-stable-structures = { workspace = true }
+ic-nns-constants = { path = "../../nns/constants" }
+itertools = { workspace = true }
+ic-nns-common = { path = "../../nns/common" }
+anyhow = { workspace = true }
 ic-cdk = { workspace = true }
+candid = { workspace = true }
+
 
 [dev-dependencies]
 ic-registry-keys = { path = "../keys" }
