+++ rs/registry/canister-client/src/data_provider.rs
@@ -67,14 +67,14 @@ impl Storable for StorableRegistryKey {
     };
 }
 
-type Memory = VirtualMemory<DefaultMemoryImpl>;
+type VM = VirtualMemory<DefaultMemoryImpl>;
 
 pub trait StableMemoryBorrower: Send + Sync {
     fn with_borrow<R>(
-        f: impl FnOnce(&StableBTreeMap<StorableRegistryKey, StorableRegistryValue, Memory>) -> R,
+        f: impl FnOnce(&StableBTreeMap<StorableRegistryKey, StorableRegistryValue, VM>) -> R,
     ) -> R;
     fn with_borrow_mut<R>(
-        f: impl FnOnce(&mut StableBTreeMap<StorableRegistryKey, StorableRegistryValue, Memory>) -> R,
+        f: impl FnOnce(&mut StableBTreeMap<StorableRegistryKey, StorableRegistryValue, VM>) -> R,
     ) -> R;
 }
 
@@ -97,8 +97,8 @@ impl<S: StableMemoryBorrower> CanisterDataProvider<S> {
         }
     }
 
-    async fn get_registry_changes_since(&self, version: u64) -> anyhow::Result<Vec<RegistryDelta>> {
-        let buff = serialize_get_changes_since_request(version)?;
+    async fn get_registry_changes_since(&self, version: u64) -> Result<Vec<RegistryDelta>, String> {
+        let buff = serialize_get_changes_since_request(version).map_err(|e| format!("{:?}", e))?;
         let response = ic_cdk::api::call::call_raw(
             Principal::from(REGISTRY_CANISTER_ID),
             "get_changes_since",
@@ -108,7 +108,8 @@ impl<S: StableMemoryBorrower> CanisterDataProvider<S> {
         .await
         .map_err(|(code, msg)| (code as i32, msg))
         .unwrap();
-        let (registry_delta, _) = deserialize_get_changes_since_response(response)?;
+        let (registry_delta, _) =
+            deserialize_get_changes_since_response(response).map_err(|e| format!("{:?}", e))?;
         Ok(registry_delta)
     }
 
@@ -116,10 +117,10 @@ impl<S: StableMemoryBorrower> CanisterDataProvider<S> {
         S::with_borrow(|local_registry| local_registry.last_key_value().map(|(k, _)| k.version))
     }
 
-    fn add_deltas(&self, deltas: Vec<RegistryDelta>) -> anyhow::Result<()> {
+    fn add_deltas(&self, deltas: Vec<RegistryDelta>) -> Result<(), String> {
         for delta in deltas {
             let string_key = std::str::from_utf8(&delta.key[..])
-                .map_err(|_| anyhow::anyhow!("Failed to convert key {:?} to string", delta.key))?;
+                .map_err(|_| format!("Failed to convert key {:?} to string", delta.key))?;
 
             if let Some(keys) = &self.keys_to_keep {
                 if keys.iter().all(|prefix| !string_key.starts_with(prefix)) {
@@ -147,7 +148,7 @@ impl<S: StableMemoryBorrower> CanisterDataProvider<S> {
     }
 
     // This function can be called in a timer to periodically update the registry stored
-    pub async fn sync_registry_stored(&self) -> anyhow::Result<()> {
+    pub async fn sync_registry_stored(&self) -> Result<(), String> {
         let mut update_registry_version = self
             .get_latest_version()
             .unwrap_or(ZERO_REGISTRY_VERSION.get());
@@ -228,4 +229,125 @@ impl<S: StableMemoryBorrower> RegistryDataProvider for CanisterDataProvider<S> {
 }
 
 #[cfg(test)]
-mod tests;
+mod tests {
+    use super::*;
+    use crate::data_provider::{
+        CanisterDataProvider, StableMemoryBorrower, StorableRegistryKey, StorableRegistryValue,
+    };
+    use ic_registry_transport::pb::v1::RegistryDelta;
+    use ic_stable_structures::memory_manager::MemoryId;
+    use ic_stable_structures::memory_manager::MemoryManager;
+    use ic_stable_structures::StableBTreeMap;
+    use std::cell::RefCell;
+    use std::collections::HashSet;
+
+    type VM = VirtualMemory<DefaultMemoryImpl>;
+
+    thread_local! {
+        static MEMORY_MANAGER: RefCell<MemoryManager<DefaultMemoryImpl>> =
+            RefCell::new(MemoryManager::init(DefaultMemoryImpl::default()));
+
+        static DUMMY_REGISTRY: RefCell<StableBTreeMap<StorableRegistryKey, StorableRegistryValue, VirtualMemory<DefaultMemoryImpl>>>  =
+        RefCell::new(StableBTreeMap::init(
+                MEMORY_MANAGER.with(|m| m.borrow().get(MemoryId::new(0)))
+        ));
+
+    }
+
+    #[derive(Default)]
+    pub struct DummyStore;
+    impl StableMemoryBorrower for DummyStore {
+        fn with_borrow<R>(
+            f: impl FnOnce(&StableBTreeMap<StorableRegistryKey, StorableRegistryValue, VM>) -> R,
+        ) -> R {
+            DUMMY_REGISTRY.with_borrow(|registry_stored| f(registry_stored))
+        }
+
+        fn with_borrow_mut<R>(
+            f: impl FnOnce(&mut StableBTreeMap<StorableRegistryKey, StorableRegistryValue, VM>) -> R,
+        ) -> R {
+            DUMMY_REGISTRY.with_borrow_mut(|registry_stored| f(registry_stored))
+        }
+    }
+
+    fn generate_deltas(num: usize) -> Vec<RegistryDelta> {
+        (1..=num)
+            .map(|i| RegistryDelta {
+                key: format!("test_key{}", i).into_bytes(),
+                values: vec![ic_registry_transport::pb::v1::RegistryValue {
+                    version: i as u64,
+                    value: format!("value{}", i).into_bytes(),
+                    deletion_marker: false,
+                }],
+            })
+            .collect()
+    }
+
+    #[test]
+    fn test_add_deltas_correctly() {
+        let provider = CanisterDataProvider::<DummyStore>::new(None);
+        let deltas = generate_deltas(10);
+
+        provider.add_deltas(deltas).unwrap();
+        let len = DUMMY_REGISTRY.with_borrow(|registry_stored| registry_stored.len());
+
+        assert_eq!(len, 10);
+        // Test `get_updates_since`
+        let updates_since = provider
+            .get_updates_since(RegistryVersion::from(5))
+            .unwrap();
+
+        // Verify that only updates with version >= 5 are returned
+        assert_eq!(updates_since.len(), 6); // 5 through 10
+
+        for (i, update) in updates_since.iter().enumerate() {
+            let expected_version = (5 + i) as u64;
+
+            assert_eq!(update.version.get(), expected_version);
+            assert_eq!(update.key, format!("test_key{}", expected_version));
+            assert_eq!(
+                update.value,
+                Some(format!("value{}", expected_version).into_bytes())
+            );
+        }
+    }
+
+    #[test]
+    fn test_add_deltas_with_keys_to_retain() {
+        let keys_to_retain = HashSet::from(["test_key1".to_string(), "test_key3".to_string()]);
+        let provider = CanisterDataProvider::<DummyStore>::new(Some(keys_to_retain));
+        let deltas = self::generate_deltas(5);
+
+        provider.add_deltas(deltas).unwrap();
+
+        let updates_since = provider
+            .get_updates_since(RegistryVersion::from(0))
+            .unwrap();
+
+        assert_eq!(updates_since.len(), 2);
+        assert_eq!(updates_since[0].version.get(), 1);
+        assert_eq!(updates_since[0].key.as_str(), "test_key1");
+        assert_eq!(
+            updates_since[0].value,
+            Some("value1".to_string().into_bytes())
+        );
+
+        assert_eq!(updates_since[1].version.get(), 3);
+        assert_eq!(updates_since[1].key.as_str(), "test_key3");
+        assert_eq!(
+            updates_since[1].value,
+            Some("value3".to_string().into_bytes())
+        );
+
+        let updates_since = provider
+            .get_updates_since(RegistryVersion::from(2))
+            .unwrap();
+
+        assert_eq!(updates_since[0].version.get(), 3);
+        assert_eq!(updates_since[0].key.as_str(), "test_key3");
+        assert_eq!(
+            updates_since[0].value,
+            Some("value3".to_string().into_bytes())
+        );
+    }
+}
