--- rs/registry/canister-data-provider/BUILD.bazel
@@ -1,43 +1,0 @@
-load("@rules_rust//rust:defs.bzl", "rust_library", "rust_test")
-
-package(default_visibility = ["//visibility:public"])
-
-DEPENDENCIES = [
-    # Keep sorted.
-    "//rs/interfaces/registry",
-    "//rs/nns/common",
-    "//rs/nns/constants",
-    "//rs/registry/transport",
-    "//rs/types/types",
-    "@crate_index//:anyhow",
-    "@crate_index//:candid",
-    "@crate_index//:ic-cdk",
-    "@crate_index//:ic-stable-structures",
-    "@crate_index//:itertools",
-]
-
-MACRO_DEPENDENCIES = []
-
-DEV_DEPENDENCIES = []
-
-MACRO_DEV_DEPENDENCIES = []
-
-ALIASES = {}
-
-rust_library(
-    name = "canister-data-provider",
-    srcs = glob(["src/**/*.rs"]),
-    aliases = ALIASES,
-    crate_name = "ic_registry_canister_data_provider",
-    proc_macro_deps = MACRO_DEPENDENCIES,
-    version = "0.9.0",
-    deps = DEPENDENCIES,
-)
-
-rust_test(
-    name = "ic_registry_canister_data_provider_test",
-    aliases = ALIASES,
-    crate = ":canister-data-provider",
-    proc_macro_deps = MACRO_DEPENDENCIES + MACRO_DEV_DEPENDENCIES,
-    deps = DEPENDENCIES + DEV_DEPENDENCIES,
-)
