--- rs/registry/canister-data-provider/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "ic-registry-canister-data-provider"
-version.workspace = true
-authors.workspace = true
-description.workspace = true
-documentation.workspace = true
-edition.workspace = true
-
-[dependencies]
-ic-interfaces-registry = { path = "../../interfaces/registry" }
-ic-registry-transport = { path = "../transport" }
-ic-types = { path = "../../types/types" }
-ic-stable-structures = { workspace = true }
-ic-nns-constants = { path = "../../nns/constants" }
-itertools = { workspace = true }
-ic-nns-common = { path = "../../nns/common" }
-anyhow = { workspace = true }
-ic-cdk = { workspace = true }
-candid = { workspace = true }
-ic-registry-canister-client = { path = "../canister-client" }
